refactor: welcome page
