+++ src/components/Welcome/Team/Team.tsx
@@ -1,6 +1,6 @@
 import { FC } from 'react';
 import styles from './team.module.scss';
-import TeamMemberComponent from './TeamMemberCoponent';
+import TeamMember from './TeamMember';
 import teamMembersData from './teamMembersData';
 
 const Team: FC = () => {
@@ -9,7 +9,7 @@ const Team: FC = () => {
       <h2>Development team</h2>
       <div className={styles.members}>
         {teamMembersData.map((teamMember, index) => (
-          <TeamMemberComponent
+          <TeamMember
             key={index}
             src={teamMember.src}
             name={teamMember.name}
