+++ src/components/Welcome/Team/TeamMember.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
-import { ITeamMember } from './interface';
+import { ITeamMember } from '../../../model/components/Welcome/Links/Team/Team';
 import styles from './team.module.scss';
 
-const TeamMemberComponent: FC<ITeamMember> = ({ src, name, description }) => {
+const TeamMember: FC<ITeamMember> = ({ src, name, description }) => {
   return (
     <div className={styles.card}>
       <img src={src} alt={name} className={styles.img} />
@@ -12,4 +12,4 @@ const TeamMemberComponent: FC<ITeamMember> = ({ src, name, description }) => {
   );
 };
 
-export default TeamMemberComponent;
+export default TeamMember;
