+++ src/components/Welcome/Team/team.module.scss
